+++ warlock_manager/apps/base_app.py
@@ -116,7 +116,7 @@ class BaseApp(ABC):
 			elif config.path:
 				# Doesn't exist, (that's fine),
 				# but the directory structure should be available to make it more simple for saving
-				self.ensure_file_parent_exists(config.path)
+				utils.ensure_file_parent_exists(config.path)
 
 	def save(self):
 		"""