+++ warlock_manager/nexus/nexus.py
@@ -1,4 +1,3 @@
-import json
 import os
 import time
 
@@ -9,6 +8,7 @@ from warlock_manager.libs.cmd import Cmd
 from warlock_manager.services.base_service import BaseService
 from warlock_manager.libs.utils import get_base_directory
 from warlock_manager.libs.logger import logger
+from warlock_manager.libs.meta import get_meta
 
 
 class Nexus:
@@ -34,11 +34,7 @@ class Nexus:
 			with open(self.email_file, 'r') as f:
 				self.email = f.read().strip()
 
-		manager_meta_path = os.path.join(get_base_directory(), '.manage.json')
-		if os.path.exists(manager_meta_path):
-			with open(manager_meta_path, 'r') as f:
-				meta = json.load(f)
-				self.game = meta['game']
+		self.game = get_meta()['game']
 
 	def set_email(self, email: str):
 		"""